+++ app/helpers/adverse_event_comments_helper.rb
@@ -1,0 +1,2 @@
+module AdverseEventCommentsHelper
+end